refactor: move TechnicalPage and test to pages directory
